refactor notification service
